refactor(channel): extract types.rs from mod.rs
